+++ tests/archive/manager/test_archive_descriptor.py
@@ -21,13 +21,17 @@ def test_archive_descriptor():
         'id': '0000',
         'name': 'My Archive',
         'description': 'This is my archive',
-        'primaryKey': ['SSN']
+        'primaryKey': ['SSN'],
+        'encoder': 'myencoder',
+        'decoder': 'mydecoder'
     }
     descriptor = ArchiveDescriptor(doc)
     assert descriptor.identifier() == '0000'
     assert descriptor.name() == 'My Archive'
     assert descriptor.description() == 'This is my archive'
     assert descriptor.primary_key() == ['SSN']
+    assert descriptor.encoder() == 'myencoder'
+    assert descriptor.decoder() == 'mydecoder'
     doc = {'id': '0001'}
     descriptor = ArchiveDescriptor(doc)
     assert descriptor.identifier() == '0001'
@@ -38,12 +42,16 @@ def test_archive_descriptor():
     descriptor = ArchiveDescriptor.create(
         name='My Archive',
         description='This is my archive',
-        primary_key='SSN'
+        primary_key='SSN',
+        encoder='myencoder',
+        decoder='mydecoder'
     )
     assert descriptor.identifier() is not None
     assert descriptor.name() == 'My Archive'
     assert descriptor.description() == 'This is my archive'
     assert descriptor.primary_key() == ['SSN']
+    assert descriptor.encoder() == 'myencoder'
+    assert descriptor.decoder() == 'mydecoder'
     descriptor = ArchiveDescriptor.create()
     assert descriptor.identifier() is not None
     assert descriptor.name() == descriptor.identifier()