+++ changelog.md
@@ -15,3 +15,5 @@
 ### 0.1.2 - ???
 
 * Proper handling of date/time objects by the default archive reader and writer
+* Optional arguments for Json encoder and decoder for persistent archives
+* Add encoder and decoder information to archive manager metadata